Minor fixes

+++ software/src/libc/arch/68k-os/unistd/stat.c
@@ -5,6 +5,6 @@
 
 int stat(const char *path, struct stat *statbuf)
 {
-	return SYSCALL2(SYS_STAT, path, (int) statbuf);
+	return SYSCALL2(SYS_STAT, (int) path, (int) statbuf);
 }
